Add files via upload
Creating a jumping jack game using an Arduino with a 16x2 LCD display and push buttons adds an exciting dimension to the project. In this game setup, the 16x2 LCD display serves as the visual interface where the game character or obstacles are presented, providing a simple yet effective way to visualize the game's progress. Push buttons are integrated to allow players to interact with the game, such as making the character jump or move in response to their button presses. The Arduino board acts as the central controller, processing the input from the push buttons and displaying the game state on the LCD screen. Overall, this project combines electronics, programming, and user interface design, making it a great learning opportunity for those looking to develop both hardware and software skills.

Components Require
- Arduino UNO
-16x2 LCD
-10k Potentiometer
-Pushbutton